Protect CheckState from misuse
To ensure consistent and protective use of the CheckState pointer that
gets passed to the Checker functions the fields have been changed to
unexported and a helpful `Update()` function has been added to manage
the correct updates to the state including time calculations.  This
should prevent any race conditions if Checker implementations store the
pointer and will also simplify the Checker implementations.